refactor: rename api to backend
